requirements.txt
fastapi
sqlalchemy
pydantic
python-dotenv
uvicorn